Remove unnecessary manifest declarations from library
Library shouldn’t specify android:allowBackup value, as doing so can
cause conflicts with applications attempting to use the library that
may have this value set to false.

Also, no need for library to specify an application label, or even
declare an application in the manifest.

Finally, without using the label, there’s no need for strings.xml in
the library.